Move all API calls into seperate instances
