+++ src/API_Implementation.h
@@ -1,8 +1,8 @@
 #ifndef API_Implementation_h
 #define API_Implementation_h
 
-// Library includes.
-#include <Callback.h>
+// Local include.
+#include "Callback.h"
 
 
 // Log messages.
@@ -10,22 +10,15 @@
 char constexpr MAX_SUBSCRIPTIONS_EXCEEDED[] = "Too many (%s) subscriptions, increase MaxSubscribtions or unsubscribe";
 #endif // !THINGSBOARD_ENABLE_DYNAMIC
 char constexpr SUBSCRIBE_TOPIC_FAILED[] = "Subscribing the given topic (%s) failed";
+// RPC data keys.
+char constexpr RPC_METHOD_KEY[] = "method";
+char constexpr RPC_PARAMS_KEY[] = "params";
+char constexpr RPC_EMPTY_PARAMS_VALUE[] = "{}";
 
 
 /// @brief Base functionality required by all API implementation
 class API_Implementation {
   public:
-    /// @brief Callback signatures
-#if THINGSBOARD_ENABLE_STL
-    using send_function = std::function<bool(char const * const topic, JsonDocument const & source, size_t const & jsonSize)>;
-    using subscribe_function = std::function<bool(char const * const topic)>;
-    using unsubscribe_function = std::function<bool(char const * const topic)>;
-#else
-    using send_function = bool (*)(char const * const topic, JsonDocument const & source, size_t const & jsonSize);
-    using subscribe_function = bool (*)(char const * const topic);
-    using unsubscribe_function = bool (*)(char const * const topic);
-#endif // THINGSBOARD_ENABLE_STL
-
     /// @brief Constructor
     API_Implementation() = default;
 
@@ -33,12 +26,23 @@ class API_Implementation {
     /// and is responsible for handling the payload and calling the appropriate previously subscribed callbacks
     /// @param topic Previously subscribed topic, we got the response over
     /// @param data Payload sent by the server over our given topic, that contains our key value pairs
-    virtual void process_response(char * const topic, JsonObjectConst const & data) const = 0;
+    virtual void Process_Response(char * const topic, JsonObjectConst const & data) const = 0;
 
     /// @brief Returns a non-owning pointer to the respone topic string, that we should have received the actual data on.
     /// Used to check, which API Implementation needs to handle the current response to a previously sent request
     /// @return Response topic null-terminated string
-    virtual const char const * get_response_topic_string() const = 0;
+    virtual const char const * Get_Response_Topic_String() const = 0;
+
+    /// @brief Forwards the call to let the API clear up any ongoing subscriptions and stop receiving information over the previously subscribed topic
+    /// @return Whether unsubscribing was successfull or not
+    virtual bool Unsubscribe_Topic() = 0;
+
+    /// @brief Forwards the call to let the API clear up any ongoing single-event subscriptions (Provision, Attribute Request, RPC Request)
+    /// and simply resubscribes the topic for all permanent subscriptions (RPC, Shared Attribute Update)
+    /// @return Whether resubscribing was successfull or not
+    virtual bool Resubscribe_Topic() {
+        return Unsubscribe_Topic();
+    }
 
 #if !THINGSBOARD_USE_ESP_TIMER
     /// @brief Internal loop method to update inernal timers for API calls that can timeout.
@@ -55,10 +59,10 @@ class API_Implementation {
     /// @param send Method which allows to send arbitrary Json payload, points to Send_Json per default
     /// @param subscribe Method which allows to subscribe to arbitrary topics, points to m_client.subscribe per default
     /// @param unsubscribe Method which allows to subscribe to arbitrary topics, points to m_client.unsubscribe per default
-    void set_client_callbacks(send_function send, subscribe_function subscribe, unsubscribe_function unsubscribe) {
-        m_send_callback = Callback(send);
-        m_subscribe_callback = Callback(subscribe);
-        m_unsubscribe_callback = Callback(unsubscribe);
+    void Set_Client_Callbacks(Callback<bool, char const * const topic, JsonDocument const & source, size_t const & jsonSize> send_callback, allback<bool, char const * const topic> subscribe_callback, Callback<bool, char const * const topic> unsubscribe_callback) {
+        m_send_callback = send_callback;
+        m_subscribe_callback = subscribe_callback;
+        m_unsubscribe_callback = unsubscribe_callback;
     }
 
   protected:
